refactor: (#14) boards 페이지 구조 분리
