+++ apps/web/src/pages/boards/index.ts
@@ -1,0 +1,1 @@
+export { BoardsPage } from "./ui/BoardsPage";
