+++ apps/web/src/pages/boards/model/types.ts
@@ -1,0 +1,18 @@
+export interface BoardPost {
+  slug: string;
+  title: string;
+  department: string;
+  timestamp: string;
+  body: string;
+}
+
+export interface BoardDirectoryItem {
+  label: string;
+  description: string;
+  posts: readonly BoardPost[];
+}
+
+export interface BoardsPageProps {
+  board: string;
+  post?: string;
+}
